add dependencies to PYTHONPATH
Currently, for a setuptools project, we add only the project itself to
PYTHONPATH. This is not enough, since a project depending on a
setuptools project needs the setuptools project on PYTHONPATH as well.

Change this to have the same behavior as LD_LIBRARY_PATH and
PKG_CONFIG_PATH, in which each dependencies gets added into the build
environment.