setuptools: fix build files left in source directory
Currently, setuptools is leaving build files in the source directory,
which is annoying and means even a force-clean doesn't clean everything.
Fix this by specifying --build-base and --egg-base. As a nice
side-effect, this means we can also get rid of the nasty site-packages
and PYTHONPATH logic, as everything appears to "just work".