init: allow non-git-repo manifests
Currently, the ws manifest is assumed to be in .repo, so you have to use
the git-repo tool. This is a nice default, but it should be overridable.
Introduce a symlink at the top of .ws pointing to the correct ws
manifest, and add a -m option to init to specify it.