ws: make -r option more flexible, change to -c
Currently, the -r option makes you point directly at a .ws directory,
which is kind of annoying. Instead, just make it run ws as if from a
different directory, and change it to -c. This brings it in line with
make -C and other tools that use a -c option for this.