conf: add ability to manually disable projects
Add ability to manually disable projects so you can satisfy them from a
distro, build them by hand, or similar. In order to do this, enable
per-project config options that are local to this workspace instance and
this project. We can build on this later to enable per-project build
arguments and similar.