add ability to manually specify per-project build args
If you want to build a particular project a little differently than what
the manifest specifies, this can be useful. For instance, you might want
to disable/enable the use of a particular library on your system.

Note that, as part of this change, I centralized the config
reading/writing to always return a singleton, and to update the config
only if it changed. This change was is needed to avoid possibly writing
to a config more than once.