make config and manifest sync up
Currently, if you change the manifest, the config.yaml file doesn't
change accordingly. This causes crashes and confusing bugs that are only
fixable by removing the .ws root and doing ws init again. Fix this so
that when you load the config, it checks against the manifest and
adds/removes projects accordingly.