env: put build dir in path, only for env cmd
In commit c5f944e9b88c4ac31ba959f011a26d567629519f, we entirely removed
the build directory from the PATH in the build env. Although this is
correct for making sure other projects don't inherit the build directory
into their env, it's still useful to have this in the PATH for the env
command. For example, when you run unit tests, you want to be able to
run them from the current source directory against data files checked
into the repository but not installed. In order to do so, you need to be
able to find the unit test executable when it's not the current
directory, and adding the build directory to the PATH has that result.