build: fix recovery bug when configure fails
When configure fails by raising an exception, currently we don't re-run
configure the next time we run (we skip straight to build instead). This
is because the build directory already exists, so needs_configure gets
set to False, and the build fails because the project hasn't been
configured.

Fix this by catching the exception, removing the build directory, and
re-raising the exception.