init: fix symlink/realpath bug
Currently, when using a manifest.xml that's deep inside a git-repo
manifest repository, we're following the manifest.xml symlink and then
doing realpath, then getting the parent directory. We actually just want
the parent of the symlink, not the parent of the file linked to by the
symlink. This is causing init failures in some configurations. Fix this.